bisenet_sd_ft/requirements.txt
albumentations>=1.4.17
numpy>=1.26.0
opencv-python-headless>=4.10.0
pandas>=2.2.3
Pillow>=10.4.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
requests>=2.32.3
scikit-image>=0.24.0
torch>=2.4.0
torchvision>=0.19.0
tqdm>=4.66.5

